backend/requirements.txt

fastapi
uvicorn
transformers
datasets
torch
pytesseract
Pillow
speechrecognition
pyaudio
openai
